Removed coerceTo("array"). Added ChatHubAsync example.
