+++ RethinkDemoWindows/RethinkDemoWindows.csproj
@@ -162,6 +162,8 @@
     <Content Include="Web.config" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ChatHub.cs" />
+    <Compile Include="ChatHubAsync.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Startup.cs" />
   </ItemGroup>
