+++ RethinkDemoWindows/Startup.cs
@@ -4,9 +4,7 @@ using Microsoft.Owin;
 using Owin;
 using Microsoft.AspNet.SignalR;
 using RethinkDb.Driver;
-using Newtonsoft.Json.Linq;
 using System.Collections.Generic;
-using RethinkDb.Driver.Net;
 
 [assembly: OwinStartup(typeof(RethinkDemoWindows.Startup))]
 
@@ -17,38 +15,6 @@ namespace RethinkDemoWindows
         public string username { get; set; }
         public string message { get; set; }
         public DateTime timestamp { get; set; }
-    }
-
-    public class ChatHub : Hub
-    {
-        public static RethinkDB r = RethinkDB.r;
-        static Connection conn;
-
-        internal static void Init()
-        {
-            conn = r.connection().connect();
-        }
-
-        public void Send(string name, string message)
-        {
-            r.db("test").table("chat")
-             .insert(new ChatMessage {
-                 username = name,
-                 message = message,
-                 timestamp = DateTime.Now
-             }).run(conn);
-        }
-
-        public JArray History(int limit)
-        {
-            var output = r.db("test").table("chat")
-                          .orderBy(r.desc("timestamp"))
-                          .limit(limit)
-                          .orderBy("timestamp")
-                          .coerceTo("array")
-                          .run<JObject>(conn);
-            return output;
-        }
     }
 
     static class ChangeHandler
@@ -75,6 +41,8 @@ namespace RethinkDemoWindows
         public void Configuration(IAppBuilder app)
         {
             ChatHub.Init();
+            //Best we can do without async void 
+            //ChatHubAsync.Init().Wait();
 
             Task.Factory.StartNew(
                 ChangeHandler.HandleUpdates,
